+++ src/splits.cpp
@@ -37,10 +37,7 @@ RawMatrix cpp_edge_to_splits(IntegerMatrix edge, IntegerVector nTip) {
 
   uintx** splits = new uintx*[n_node];
   for (uintx i = 0; i != n_node; i++) {
-    splits[i] = new uintx[n_bin];
-    for (uintx j = 0; j != n_bin; j++) {
-      splits[i][j] = 0;
-    }
+    splits[i] = new uintx[n_bin](); // () zero-initializes
   }
 
   for (uintx i = 0; i != n_tip; i++) {