+++ NEWS.md
@@ -16,6 +16,8 @@
 ## Enhancements
 
 - Improve support for unrooted trees in `as.Splits()`.
+- Use methods so `Reorder` functions can handle `multiPhylo` objects and edges.
+  Deprecates `PostorderEdges()`.
 - Handle funny node orders.
 - Support continuous characters in `ReadCharacters()`.
 - Improve performance of `as.logical.Splits()` and related functions.